[issue-344] - Extend MicroProfile OpenAPI tests to cover the multiple deployment scenario
