+++ microprofile-open-api/src/test/java/org/jboss/eap/qe/microprofile/openapi/model/AnotherOpenApiModelReader.java
@@ -21,7 +21,7 @@ public class AnotherOpenApiModelReader implements OASModelReader {
         Info info = OASFactory.createInfo();
         String newDescription = ((info.getDescription() == null) || (info.getDescription().isEmpty()) ? ""
                 : info.getDescription()) +
-                String.format("[Generated by router: %s]", OpenApiFilter.LOCAL_TEST_ROUTER_FQDN);
+                String.format("[Generated by router: %s]", AnotherOpenApiFilter.ANOTEHR_LOCAL_TEST_ROUTER_FQDN);
         info.setDescription(newDescription);
         api.setInfo(info);
         return api;
